requirements.txt
annotated-types==0.7.0
anyio==4.10.0
beautifulsoup4==4.13.5
cachetools==5.5.2
certifi==2025.8.3
cfgv==3.4.0
charset-normalizer==3.4.3
distlib==0.4.0
dnspython==2.8.0
filelock==3.19.1
google-auth==2.40.3
google-genai==1.38.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.14
idna==3.10
nodeenv==1.9.1
platformdirs==4.4.0
pre_commit==4.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
pymongo==4.15.1
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.5
rsa==4.9.1
sentry-sdk==2.39.0
sniffio==1.3.1
soupsieve==2.8
tenacity==9.1.2
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
virtualenv==20.34.0
websockets==15.0.1
